size backend corrections

+++ migrations/20210404092232-filesize.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    await queryInterface.addColumn('Events', 'maxFileSize', { type: Sequelize.INTEGER(12) });
+    await queryInterface.addColumn('Events', 'maxFileSize', { type: Sequelize.BIGINT(20) });
   },
 
   down: async (queryInterface, Sequelize) => {
